chat: move ConvSource interface into libkb
