+++ go/libkb/globals.go
@@ -86,6 +86,8 @@ type GlobalContext struct {
 
 	CardCache *UserCardCache // cache of keybase1.UserCard objects
 
+	ConvSource ConversationSource // source of remote message bodies for chat
+
 	// Can be overloaded by tests to get an improvement in performance
 	NewTriplesec func(pw []byte, salt []byte) (Triplesec, error)
 }
